Опция 'prebuild' ; Разнесено на компоненты ; Добавлена директория лдя тестов.

+++ src/scripts/cc.js
@@ -1,4 +1,15 @@
 #!/usr/bin/node
+/*
+ * compiler <https://github.com/1024sparrow/compiler>
+ * Утилита для многоэтапной сборки проектов.
+ * Utility for multistage project building.
+ * 
+ * Авторское право (c) 2017, Борис Васильев.
+ * Публикуется под лицензией MIT.
+ *
+ * Copyright (c) 2017, Boris Vasilyev.
+ * Released under the MIT license.
+ */
 
 var program = require('commander');
 var path = require('path');
@@ -6,10 +17,14 @@ var fs = require('fs');
 var fse = require('fs-extra');
 var child_process = require('child_process');
 var tmpFile = require('tmp').fileSync();
+var StringDecoder = require('string_decoder').StringDecoder;
+var DECODER = new StringDecoder('utf8');
+
+
 program.version('1.0')
     .option('-i, --info', 'Показать информацию по предстоящей сборке');
 program.on('--help', function(){
-    console.log('\nКомпилятор кода.\n================\nПрограмма принимает в качестве параметра путь к compile.ini.\nИмя файла \'compile.ini\' упоминается условно - вы можете произвольным образом назвать этот файл. Я, например, называю его как \'<название_проекта>.pro\'.\nВ папках с исходниками предполагается наличие файлов __meta__ (если такого файла в папке нет, то компилироваться папка не будет)\ncompile.ini:\n------------\nЗдесь определяются правила компиляции (строковый идентификатор и обрабатывающая функция(на JavaScript))\nОформляется как NodeJS-модуль:\nmodule.exports = {\n    target: \'compiled\', // <-- здесь будет формироваться результат компиляции\n    file_script_dir: \'\', //* относительный путь к директории, относительно которой будут указываться пути к скриптам для обработки файлов. Если свойство не указано - пути относительно директории, где находится файл compile.ini.\n    file: {\n        css:function(srcText){ // <-- типы файлов \'css\' будут обрабатываться вот этой функцией. Функция должна вернуть результат в виде строки.\n        }\n        // вместо функции можно указать путь к скрипту, который должен преобразовать данные в файле, полный путь к которому будет передан единственным параметром.\n    },\n    dir_script_dir: \'\', //* относительный путь к директории, относительно которой будут указываться пути к скриптам для обработки директорий. Если свойство не указано - пути относительно директории, где находится файл compile.ini.\n    dir: {\n        tab_app:function(dirPath, dirName){ // <--передаётся абсолютный путь до папки, содержащей целевую папку, и имя целевой папки. Этот обработчик будет применяться к директориям, помеченным как \'tab_app\'.\n        }\n        // вместо функции можно указать путь к скрипту, который должен преобразовать директорию. Скрипту передаются два параметра - такие, как передавались бы в фунцию (см. выше).\n    }\n}\n\n__meta__:\n---------\nЗдесь определяются цели компиляции, исходники и указываются идентификаторы обработчиков (из compile.ini).\nОформляется как JSON.\n{\n    files: [{..},{..}], //<-- Если это свойство есть, то собираем указанные файлы. Если этого свойства нет, то тупо копируем всю директорию.\n    dir_proc: [\'..\',\'..\'] //<-- Перечень обработчиков, которые необходимо применить к результирующей директории. Если нужно сохранить директорию (т.е. результаты компиляции будут в такой же папке, а не положены вместо неё), то свойство должно быть, пусть в массиве и не будет элементов.\n}\n\nФормат описания правила компиляции файла (такой объект кладём в массив files):\n{\n    target: \'..\',//имя, не путь\n    type: [\'..\',\'..\'],//* обработчики файловые (текстовые идентификаторы из compile.ini), которые нужно применить (постобработка, после формирования из составляющих)\n    source:{\n        list: [\'1.js\', \'2.js\'],\n        template: \'..\', //* путь до файла с шаблоном\n        types:{\n            \'1.js\': [\'..\',\'..\'] // обработчики файловые (текстовые идентификаторы из compile.ini), которые нужно применить (предобработка, перед вставкой в целевой файл)\n        }\n    }\n}\n// * - необязательное свойство\n\nПри написании шаблона как ссылаться на файлы-исходники:\n{%% 1.js %%}\n1.js - имя файла (не забудьте его прописать в __meta__ !). Между \'%\' и именем файла обязательно должен быть один пробел.\n\nПростейший файл __meta__:\n-------------------------\n{\n    \"dir_proc\": []\n}\nЧто он делает - сохраняет директорию как есть. Если файл __meta__ отсутсвует или в нём нет свойства \'dir_proc\', такой папки в результатах компиляции не будет.\n\nСм. также:\n    - https://www.npmjs.com/package/node-minify\n    - var stripComments = require(\'strip-comments\');data = stripComments(data);\nАвтор: Васильев Б.П.\n');
+    console.log('\nКомпилятор кода.\n================\nПрограмма принимает в качестве параметра путь к compile.ini.\nИмя файла \'compile.ini\' упоминается условно - вы можете произвольным образом назвать этот файл. Я, например, называю его как \'<название_проекта>.pro\'.\nВ папках с исходниками предполагается наличие файлов __meta__ (если такого файла в папке нет, то компилироваться папка не будет)\ncompile.ini:\n------------\nЗдесь определяются правила компиляции (строковый идентификатор и обрабатывающая функция(на JavaScript))\nОформляется как NodeJS-модуль:\nmodule.exports = {\n    target: \'compiled\', // <-- здесь будет формироваться результат компиляции\n    file_script_dir: \'\', //* относительный путь к директории, относительно которой будут указываться пути к скриптам для обработки файлов. Если свойство не указано - пути относительно директории, где находится файл compile.ini.\n    file: {\n        css:function(srcText){ // <-- типы файлов \'css\' будут обрабатываться вот этой функцией. Функция должна вернуть результат в виде строки.\n        }\n        // вместо функции можно указать путь к скрипту, который должен преобразовать данные в файле, полный путь к которому будет передан единственным параметром.\n    },\n    dir_script_dir: \'\', //* относительный путь к директории, относительно которой будут указываться пути к скриптам для обработки директорий. Если свойство не указано - пути относительно директории, где находится файл compile.ini.\n    dir: {\n        tab_app:function(dirPath, dirName){ // <--передаётся абсолютный путь до папки, содержащей целевую папку, и имя целевой папки. Этот обработчик будет применяться к директориям, помеченным как \'tab_app\'.\n        }\n        // вместо функции можно указать путь к скрипту, который должен преобразовать директорию. Скрипту передаются два параметра - такие, как передавались бы в фунцию (см. выше).\n    }\n}\n\n__meta__:\n---------\nЗдесь определяются цели компиляции, исходники и указываются идентификаторы обработчиков (из compile.ini).\nОформляется как JSON.\n{\n    files: [{..},{..}], //<-- Если это свойство есть, то собираем указанные файлы. Если этого свойства нет, то тупо копируем всю директорию.\n    dir_proc: [\'..\',\'..\'], //<-- Перечень обработчиков, которые необходимо применить к результирующей директории. Если нужно сохранить директорию (т.е. результаты компиляции будут в такой же папке, а не положены вместо неё), то свойство должно быть, пусть в массиве и не будет элементов.\n    remove: [] //<-- какие файлы и директории (от компиляции поддиректорий) надо удалить. Удалено будет перед запуском обработчиков из dir_proc\n}\n\nФормат описания правила компиляции файла (такой объект кладём в массив files):\n{\n    target: \'..\',//имя, не путь\n    type: [\'..\',\'..\'],//* обработчики файловые (текстовые идентификаторы из compile.ini), которые нужно применить (постобработка, после формирования из составляющих)\n    source:{\n        list: [\'1.js\', \'2.js\'],\n        template: \'..\', //* путь до файла с шаблоном\n        types:{\n            \'1.js\': [\'..\',\'..\'] // обработчики файловые (текстовые идентификаторы из compile.ini), которые нужно применить (предобработка, перед вставкой в целевой файл)\n        }\n    }\n}\n// * - необязательное свойство\n\nПри написании шаблона как ссылаться на файлы-исходники:\n{%% 1.js %%}\n1.js - имя файла (не забудьте его прописать в __meta__ !). Между \'%\' и именем файла обязательно должен быть один пробел.\n\nПростейший файл __meta__:\n-------------------------\n{\n    \"dir_proc\": []\n}\nЧто он делает - сохраняет директорию как есть. Если файл __meta__ отсутсвует или в нём нет свойства \'dir_proc\', такой папки в результатах компиляции не будет.\n\nСм. также:\n    - https://www.npmjs.com/package/node-minify\n    - var stripComments = require(\'strip-comments\');data = stripComments(data);\nАвтор: Васильев Б.П.\n');
 });
 program.on('--info', function(){
     console.log('123');
@@ -27,13 +42,29 @@ program.action(function(compileIniPath){
     }
     const destPathStart = processor.target;
     fse.removeSync(path.dirname(compile_ini_path) + '/' + destPathStart);
-    var stack = [path.dirname(compile_ini_path)];
-    console.log(stack);
+    var compileIniDir = path.dirname(compile_ini_path);
+    console.log('\033[93mНачинаю сборку исходников в директории \''+compileIniDir+'\'\033[0m');
+    var stack = [compileIniDir];
     var dirStack = [];
     while (stack.length){
         var parent = stack.pop();
-        if (fs.existsSync(parent + '/__meta__'))
+        tmp = parent + '/__meta__';
+        if (fs.existsSync(tmp)){
             dirStack.push(parent);
+            
+            
+            try{
+                var meta = JSON.parse(fs.readFileSync(tmp, 'utf8'));
+            } catch(e) {
+                console.log('Файл \''+tmp+'\' не является корректным JSON-файлом. Операция компиляции прервана.');
+                console.log('Описание ошибки: '+e);
+                return; 
+            }
+            if ((!meta.hasOwnProperty('files')) && (tmp != path.resolve(compileIniDir, '__meta__')))
+                continue;
+        }
+
+
         var children = fs.readdirSync(parent);
         for (var i = 0 ; i < children.length ; i++){
             tmp = parent + '/' + children[i];
@@ -46,23 +77,27 @@ program.action(function(compileIniPath){
     var destDir;
     while (dirStack.length){
         var dirCandidate = dirStack.pop();
+        console.log('\033[91mОбрабатываю директорию \''+path.relative(compileIniDir, dirCandidate)+'/\'\033[0m');
         tmp = dirCandidate+'/__meta__';
         try{
             var meta = JSON.parse(fs.readFileSync(tmp, 'utf8'));
-        } catch(err) {
+        } catch(e) {
             console.log('Файл \''+tmp+'\' не является корректным JSON-файлом. Операция компиляции прервана.');
+            console.log('Описание ошибки: '+e);
             break;
         }
-        t = path.dirname(compile_ini_path);
+        t = compileIniDir;
         tmp =  path.relative(t, dirCandidate);
         tmp = path.resolve(t, destPathStart, tmp);
         tmp = tmp.replace(/\/+/g, '/')
                  .replace(/(\/)$/, '');
-        if (!applyMeta(meta, dirCandidate, tmp, processor, path.dirname(compile_ini_path))){
+        if (!applyMeta(meta, dirCandidate, tmp, processor, compileIniDir)){
             console.log('Операция компиляции прервана.');
-            break;
+            return;
         }
     }
+    
+    console.log('\033[93mСборка успешно завершена\033[0m');
 });
 function applyMeta(meta, srcPath, destPath, processor, processorDirPath){ 
     const tmpDestPath = destPath + '.tmp';
@@ -72,12 +107,27 @@ function applyMeta(meta, srcPath, destPath, processor, processorDirPath){
     if (bF){
         for (const file of meta.files){
             const hasTempl = file.source.hasOwnProperty('template');
-            let retval = hasTempl ? fs.readFileSync(srcPath + '/'  + file.source.template, 'utf8') : '';
+            let retval = '';
+            if (hasTempl){
+                const tmp = path.resolve(destPath, file.source.template);
+                if (fs.existsSync(tmp))
+                    retval = fs.readFileSync(tmp, 'utf8');
+                else{
+                    const t = path.resolve(srcPath, file.source.template);
+                    if (fs.existsSync(t)){
+                        retval = fs.readFileSync(t, 'utf8');
+                    }
+                    else{
+                        console.log(`Файл шаблона ${tmp} не найден. Операция компиляции прервана.`);
+                        return false;
+                    }
+                }
+            }
             for (const srcFile of file.source.list){
                 let tmp = destPath + '/'  + srcFile;
                 if (fs.existsSync(tmp)){
                     tmp = fs.readFileSync(tmp, 'utf8');
-                    //мы использовали в качестве исходников результаты компиляции поддиректории - мы должны будем перед окончанием компиляции текущей директории удалить эту директорию с исходниками (это надо делать ДО того, как будут мёржиться временная директория в целевую)
+                    
                 }
                 else{
                     tmp = srcPath + '/'  + srcFile;
@@ -95,13 +145,12 @@ function applyMeta(meta, srcPath, destPath, processor, processorDirPath){
                             const tmpType = typeof tmpFunc;
                             if (tmpType === 'string'){
                                 fs.writeFileSync(tmpFile.name, tmp, 'utf8');
-                                console.log(
-                                        child_process.execSync(
-                                            path.resolve(processorDirPath, tmpFunc) + ' ' + tmpFile.name
-                                            )
-                                        );
+                                const tt = path.resolve(processorDirPath, tmpFunc) + ' ' + tmpFile.name;
+                                const ou = child_process.execSync(tt, {encoding:'utf8', stdio:[0,1,2]});
+                                if (ou)
+                                    console.log(DECODER.write(ou));
                                 tmp = fs.readFileSync(tmpFile.name, 'utf8');
-                                //tmpFile.cleanupCallback();
+                                
                             }
                             else if (tmpType === 'function'){
                                 tmp = tmpFunc(tmp);
@@ -121,7 +170,7 @@ function applyMeta(meta, srcPath, destPath, processor, processorDirPath){
                 else
                     retval += tmp;
             }
-            //Применяем обработчики к получившемуся файлу
+            
             if (file.hasOwnProperty('type')){
                 for (const filetype of file.type){
                     if (processor.file.hasOwnProperty(filetype)){
@@ -129,14 +178,15 @@ function applyMeta(meta, srcPath, destPath, processor, processorDirPath){
                         const tmpType = typeof tmpFunc;
                         if (tmpType === 'string'){
                             fs.writeFileSync(tmpFile.name, retval, 'utf8');
-                            console.log(
+                            const ou = 
                                     child_process.execSync(
-                                        path.resolve(processorDirPath, tmpFunc) + ' ' + tmpFile.name
-                                        )
-                                    .toString()
-                                    );
+                                        path.resolve(processorDirPath, tmpFunc) + ' ' + tmpFile.name,
+                                        {encoding:'utf8', stdio:[0,1,2]}
+                                        );
+                            if (ou)
+                                console.log(DECODER.write(ou));
                             retval = fs.readFileSync(tmpFile.name, 'utf8');
-                            //tmpFile.cleanupCallback();
+                            
                         }
                         else if (tmpType === 'function'){
                             retval = tmpFunc(retval);
@@ -156,8 +206,24 @@ function applyMeta(meta, srcPath, destPath, processor, processorDirPath){
         }
     }
     else{
-        copyDirContent(srcPath, tmpDestPath);
-        fs.unlinkSync(`${tmpDestPath}/__meta__`);
+        if (srcPath === processorDirPath){
+            copyDirContent(destPath, tmpDestPath);
+        }
+        else{
+            copyDirContent(srcPath, tmpDestPath);
+            fs.unlinkSync(`${tmpDestPath}/__meta__`);
+        }
+    }
+    if (meta.hasOwnProperty('remove')){
+        for (const i of meta.remove){
+            const t = path.resolve(destPath, i);
+            if (fs.existsSync)
+                fse.removeSync(t);
+            else{
+                console.log(`Невозможно удалить файл/директорию '${t}' - такого нет. Правила компиляции некорректны. Операция компиляции прервана.`);
+                return false;
+            }
+        }
     }
     if (bD){
         if (!mergeDirs(tmpDestPath, destPath)){
@@ -171,11 +237,13 @@ function applyMeta(meta, srcPath, destPath, processor, processorDirPath){
                 const p1 = path.dirname(destPath);
                 const p2 = destPath.match(/([^\/])*$/g)[0];
                 if (tmpType === 'string'){
-                    console.log(
+                    console.log(`Выполняется скрипт '${path.resolve(processorDirPath, tmpFunc)} ${p1} ${p2}'.`);
+                    const ou = 
                             child_process.execSync(
                                 `${path.resolve(processorDirPath, tmpFunc)} ${p1} ${p2}`
-                                )
-                            );
+                                );
+                    if (ou)
+                        console.log(DECODER.write(ou));
                 }
                 else if (tmpType === 'function'){
                     tmpFunc(p1, p2);
@@ -184,6 +252,9 @@ function applyMeta(meta, srcPath, destPath, processor, processorDirPath){
                     console.log(`Для типа директории '${dirFuncId}' указан некорректный обработчик: опреация компиляции прервана.`);
                     return false;
                 }
+            }
+            else{
+                console.log(`Обработчик диретории '${dirFuncId}' не найден - директория '${destPath}' осталась необработанной.`);
             }
         }
     }
@@ -209,23 +280,7 @@ function createFullPath(fullPath){
     }
 }
 function copyDirContent(srcDirPath, destDirPath){
-    var stack = [[srcDirPath, destDirPath]];
-    while (stack.length){
-        const [parentSrc, parentDest] = stack.pop();
-        const children = fs.readdirSync(parentSrc);
-        for (const i of children){
-            const tmpSrc = `${parentSrc}/${i}`;
-            const tmpDest = `${parentDest}/${i}`;
-            const stat = fs.statSync(tmpSrc);
-            if (stat.isDirectory()){
-                fs.mkdirSync(tmpDest);
-                stack.push([tmpSrc, tmpDest]);
-            }
-            else{
-                fs.createReadStream(tmpSrc).pipe(fs.createWriteStream(tmpDest));
-            }
-        }
-    }
+    fse.copySync(srcDirPath, destDirPath, {dereference:true});
 }
 function mergeDirs(p_fromDir, p_toDir){
 
@@ -278,21 +333,17 @@ function mergeDirs(p_fromDir, p_toDir){
     while (list.length){
         const a = list.pop();
         const aTo = path.resolve(p2, a);
-        if (fs.existsSync(aTo)){
-            const aToStat = fs.fs.statSync(aTo);
-            const wordExists = aToStat.isDirectory() ? 'директория' : 'файл';
-            console.log(`Невозможно записать файл '${aTo}' - уже есть ${wordExists} с таким именем.`);
-            return false;
-        }
+        
         const aFrom = path.resolve(p1, a);
         fs.renameSync(aFrom, aTo);
     }
     while (dirList.length){
-        const a = dirPost.pop();
+        const a = dirList.pop();
         fs.rmdirSync(path.resolve(p1, a));
     }
     fs.rmdirSync(p1);
     return true;
 }
 program.parse(process.argv);
+
